Fully rewrite Fade mode to use a temporary render buffer
Instead of doing some ugly hacks to all kinds of places that never
fully worked, we instead now render the hotbar items to a custom
temporary buffer. This buffer can then be rendered as a texture
with transparency to the original framebuffer.

This also removes the requirement for a custom set of mixins, the
one for a restart after changing the animation mode, and should also
massively increase mod compatibility.